+++ R/grabMzxmlFunctions.R
@@ -16,12 +16,16 @@
 #' @param grab_what What data should be read from the file? Options include
 #'   "MS1" for data only from the first spectrometer, "MS2" for fragmentation
 #'   data, "BPC" for rapid access to the base peak chromatogram, and "TIC" for
-#'   rapid access to the total ion chromatogram. These options can be combined
-#'   (i.e. `grab_data=c("MS1", "MS2", "BPC")`) or this argument can be set to
-#'   "everything" to extract all of the above. Option "EIC" is useful when
-#'   working with files whose total size exceeds working memory - it first
-#'   extracts all relevant MS1 and MS2 data, then discards data outside of the
-#'   mass range(s) calculated from the provided mz and ppm.
+#'   rapid access to the total ion chromatogram. DAD and chromatogram ("DAD" and
+#'   "chroms") are unavailable for mzXML files. Metadata can be accessed with
+#'   "metadata", which provides information about the instrument and time the
+#'   file was run. These options can be combined (i.e. `grab_data=c("MS1",
+#'   "MS2", "BPC")`) or this argument can be set to "everything" to extract all
+#'   of the above. Options "EIC" and "EIC_MS2" are useful when working with
+#'   files whose total size exceeds working memory - it first extracts all
+#'   relevant MS1 and MS2 data, respectively, then discards data outside of the
+#'   mass range(s) calculated from the provided mz and ppm. The default,
+#'   "everything", includes all MS1, MS2, BPC, TIC, and metadata.
 #' @param verbosity Three levels of processing output to the R console are
 #'   available, with increasing verbosity corresponding to higher integers. A
 #'   verbosity of zero means that no output will be produced, useful when
@@ -36,21 +40,33 @@
 #'   combined with `grab_what = "EIC"` (see above).
 #' @param rtrange Not supported for mzXML data. Only provided here so as to
 #'   throw a friendly warning rather than an unexpected error.
+#' @param prefilter A single number corresponding to the minimum intensity of
+#'   interest in the MS1 data. Data points with intensities below this threshold
+#'   will be silently dropped, which can dramatically reduce the size of the
+#'   final object. Currently only works with MS1 data, but could be expanded
+#'   easily to handle more.
 #'
 #' @return A list of `data.table`s, each named after the arguments requested in
-#'   grab_what. $MS1 contains MS1 information, $MS2 contains fragmentation info,
-#'   etc. MS1 data has three columns: retention time (rt), mass-to-charge (mz),
-#'   and intensity (int). MS2 data has five: retention time (rt), precursor m/z
-#'   (premz), fragment m/z (fragmz), fragment intensity (int), and collision
-#'   energy (voltage). Data requested that does not exist in the provided files
-#'   (such as MS2 data requested from MS1-only files) will return an empty
-#'   (length zero) data.table.
+#'   grab_what. E.g. $MS1 contains MS1 information, $MS2 contains fragmentation
+#'   info, etc. MS1 data has four columns: retention time (rt), mass-to-charge
+#'   (mz), intensity (int), and filename. MS2 data has six: retention time (rt),
+#'   precursor m/z (premz), fragment m/z (fragmz), fragment intensity (int),
+#'   collision energy (voltage), and filename. Data requested that does not
+#'   exist in the provided files (such as MS2 data requested from MS1-only
+#'   files) will return an empty (length zero) data.table. The data.tables
+#'   extracted from each of the individual files are collected into one large
+#'   table using data.table's `rbindlist`. $metadata is a little weirder because
+#'   the metadata doesn't fit neatly into a tidy format but things are hopefully
+#'   named helpfully. Data requested that does not exist in the provided files
+#'   (such as DAD or chromatogram data) will return an empty (zero-row)
+#'   data.table.
 #'
 #' @export
 #'
 #' @examples
 #' sample_file <- system.file("extdata", "LB12HL_AB.mzXML.gz", package = "RaMS")
 #' file_data <- grabMzxmlData(sample_file, grab_what="MS1")
+#' \dontrun{
 #' # Extract MS1 data and a base peak chromatogram
 #' file_data <- grabMzxmlData(sample_file, grab_what=c("MS1", "BPC"))
 #' # Extract EIC for a specific mass
@@ -62,9 +78,10 @@
 #' # Extract MS2 data
 #' sample_file <- system.file("extdata", "DDApos_2.mzXML.gz", package = "RaMS")
 #' MS2_data <- grabMzxmlData(sample_file, grab_what="MS2")
+#' }
 #'
 grabMzxmlData <- function(filename, grab_what, verbosity=0,
-                          rtrange=NULL, mz=NULL, ppm=NULL){
+                          rtrange=NULL, mz=NULL, ppm=NULL, prefilter=-1){
   if(verbosity>1){
     cat(paste0("\nReading file ", basename(filename), "... "))
     last_time <- Sys.time()
@@ -73,23 +90,28 @@ grabMzxmlData <- function(filename, grab_what, verbosity=0,
 
   checkFileType(xml_data, "mzXML")
   rtrange <- checkRTrange(rtrange)
-  file_metadata <- grabMzxmlEncodingData(xml_data)
+  prefilter <- checkProvidedPrefilter(prefilter)
 
   output_data <- list()
 
   if("everything"%in%grab_what){
-    if(length(setdiff(grab_what, "everything"))&&verbosity>0){
+    extra_grabs <- setdiff(grab_what, "everything")
+    if(any(c("MS1", "MS2", "BPC", "TIC", "metadata")%in%extra_grabs)&&verbosity>0){
       message(paste("Heads-up: grab_what = `everything` includes",
-                    "MS1, MS2, BPC, and TIC data"))
-      message("Ignoring additional grabs")
+                    "MS1, MS2, BPC, TIC, and meta data"))
+      message("Ignoring duplicate specification")
     }
-    grab_what <- c("MS1", "MS2", "BPC", "TIC", "metadata")
+    grab_what <- unique(c("MS1", "MS2", "BPC", "TIC", "metadata", extra_grabs))
+  }
+
+  if(any(c("MS1", "MS2", "EIC", "EIC_MS2")%in%grab_what)){
+    file_metadata <- grabMzxmlEncodingData(xml_data)
   }
 
   if("MS1"%in%grab_what){
     if(verbosity>1)last_time <- timeReport(last_time, text = "Reading MS1 data...")
     output_data$MS1 <- grabMzxmlMS1(xml_data = xml_data, rtrange=rtrange,
-                                    file_metadata = file_metadata)
+                                    file_metadata = file_metadata, prefilter = prefilter)
   }
 
   if("MS2"%in%grab_what){
@@ -113,8 +135,8 @@ grabMzxmlData <- function(filename, grab_what, verbosity=0,
     checkProvidedMzPpm(mz, ppm)
     if(verbosity>1)last_time <- timeReport(last_time, text = "Extracting EIC...")
     if(!"MS1"%in%grab_what){
-      init_dt <- grabMzxmlMS1(xml_data=xml_data, file_metadata = file_metadata,
-                              rtrange=rtrange)
+      init_dt <- grabMzxmlMS1(xml_data = xml_data, file_metadata = file_metadata,
+                              rtrange = rtrange, prefilter = prefilter)
     } else {
       init_dt <- output_data$MS1
       if(!nrow(init_dt))stop("Something weird - can't find MS1 data to subset")
@@ -141,6 +163,28 @@ grabMzxmlData <- function(filename, grab_what, verbosity=0,
       init_dt[premz%between%pmppm(mass = mass, ppm = ppm)]
     })
     output_data$EIC_MS2 <- rbindlist(EIC_MS2_list)
+  }
+
+  if("chroms" %in% grab_what){
+    warning(paste("grab_what = 'chroms' not available for mzXML documents,",
+                  "returning empty table"))
+    output_data$chroms <- data.table(
+      chrom_type = character(0),
+      chrom_index = character(0),
+      target_mz = numeric(0),
+      product_mz = numeric(0),
+      int = numeric(0),
+      filename = character(0))
+  }
+
+  if("DAD" %in% grab_what){
+    warning(paste("grab_what = 'DAD' not available for mzXML documents,",
+                  "returning empty table"))
+    output_data$chroms <- data.table(
+      rt = numeric(0),
+      lambda = numeric(0),
+      int = numeric(0),
+      filename = character(0))
   }
 
   if("metadata"%in%grab_what){
@@ -187,27 +231,59 @@ grabMzxmlMetadata <- function(xml_data){
     names(inst_vals) <- "Instrument data"
   }
 
-  mslevel_nodes <- xml2::xml_find_all(xml_data, xpath = "//d1:scan")
-  if(length(mslevel_nodes)>0){
-    mslevels <- paste0("MS", unique(xml2::xml_attr(mslevel_nodes, "msLevel")),
-                       collapse = ", ")
-  } else {
-    mslevels <- "None found"
-  }
+  scan_nodes <- xml2::xml_find_all(xml_data, xpath = "//d1:scan")
 
-  polarity_nodes <- xml2::xml_find_all(xml_data, xpath = "//d1:scan")
-  if(length(polarity_nodes)>0){
-    polarities <- paste0(unique(xml2::xml_attr(polarity_nodes, "polarity")),
-                         collapse = ", ")
-  } else {
-    polarities <- "None found"
-  }
+  n_spectra <- length(scan_nodes)
 
+  if(n_spectra>0){
+    centroided <- as.integer(unique(xml2::xml_attr(scan_nodes, "centroided")))
+    # 1 if centroided, 0 if not (i.e. profile mode)
+    if (1 %in% centroided) {
+      centroided <- TRUE
+    } else {
+      centroided <- FALSE
+    }
+
+    ms_levels <- paste0("MS", unique(xml2::xml_attr(scan_nodes, "msLevel")),
+                        collapse = ", ")
+
+    mz_lowest <- min(as.numeric(xml2::xml_attr(scan_nodes, "lowMz")))
+
+    mz_highest <- max(as.numeric(xml2::xml_attr(scan_nodes, "highMz")))
+
+    rt <- xml2::xml_attr(scan_nodes, "retentionTime")
+    unit <- unique(gsub(".*[0-9]", "", rt))
+    rt <- gsub("[^0-9.-]", "", rt)
+    rt <- as.numeric(rt)
+    if("S"%in%unit) rt <- rt/60
+
+    rt_start <- min(rt)
+    rt_end <- max(rt)
+
+    polarities <- unique(xml2::xml_attr(scan_nodes, "polarity"))
+    polarities[polarities %in% "+"] <- "positive"
+    polarities[polarities %in% "-"] <- "negative"
+
+  } else {
+    centroided <- NA
+    ms_levels <- NA_integer_
+    mz_lowest <- NA_real_
+    mz_highest <- NA_real_
+    rt_start <- NA_real_
+    rt_end <- NA_real_
+    polarities <- NA_character_
+  }
 
   metadata <- data.table(
     source_file=source_file,
     inst_data=list(inst_vals),
-    mslevels=mslevels,
+    n_spectra=n_spectra,
+    ms_levels=ms_levels,
+    mz_lowest=mz_lowest,
+    mz_highest=mz_highest,
+    rt_start=rt_start,
+    rt_end=rt_end,
+    centroided=centroided,
     polarity=polarities
   )
 }
@@ -249,10 +325,12 @@ grabMzxmlEncodingData <- function(xml_data){
 #'   retention times of interest. Providing a range here can speed up load times
 #'   (although not enormously, as the entire file must still be read) and reduce
 #'   the final object's size.
+#' @param prefilter The lowest intensity value of interest, used to reduce file
+#'   size (and especially useful for profile mode data with many 0 values)
 #'
 #' @return A `data.table` with columns for retention time (rt), m/z (mz),
 #' and intensity (int).
-grabMzxmlMS1 <- function(xml_data, file_metadata, rtrange){
+grabMzxmlMS1 <- function(xml_data, file_metadata, rtrange, prefilter){
   ms1_xpath <- '//d1:scan[@msLevel="1" and @peaksCount>0]'
   ms1_nodes <- xml2::xml_find_all(xml_data, ms1_xpath)
 
@@ -267,7 +345,8 @@ grabMzxmlMS1 <- function(xml_data, file_metadata, rtrange){
   dt_data <- mapply(cbind, rt_vals, mz_int_vals, SIMPLIFY = FALSE)
   dt <- as.data.table(do.call(what=rbind, dt_data))
   names(dt) <- c("rt", "mz", "int")
-  dt
+  int <- NULL #To prevent R CMD check "notes" when using data.table syntax
+  dt[int>prefilter]
 }
 
 
@@ -330,7 +409,7 @@ grabMzxmlMS2 <- function(xml_data, file_metadata, rtrange){
 #' @return A `data.table` with columns for retention time (rt), and intensity (int).
 grabMzxmlBPC <- function(xml_data, TIC=FALSE, rtrange){
   scan_nodes <- xml2::xml_find_all(
-    xml_data, '//d1:scan[@msLevel="1" and @peaksCount>0]'
+    xml_data, '//d1:scan[@msLevel="1"]'
   )
   rt_chrs <- xml2::xml_attr(scan_nodes, "retentionTime")
   rt_vals <- as.numeric(gsub(pattern = "PT|S", replacement = "", rt_chrs))
@@ -361,12 +440,16 @@ grabMzxmlBPC <- function(xml_data, TIC=FALSE, rtrange){
 
 grabMzxmlSpectraRt <- function(xml_nodes){
   rt_attrs <- xml2::xml_attr(xml_nodes, "retentionTime")
+  rt_unit <- unique(gsub(".*[0-9]", "", rt_attrs))
   rt_vals <- as.numeric(gsub("PT|S", "", rt_attrs))
-  if(any(rt_vals>150)){
-    # Guess RT is in seconds if the run is more than 150 long
-    # A 2.5 minute run is unheard of, and a 2.5 hour run is unheard of
-    rt_vals <- rt_vals/60
-  }
+
+  if ("S" %in% rt_unit) rt_vals <- rt_vals/60
+  # if(any(rt_vals>150)){
+  #   # Guess RT is in seconds if the run is more than 150 long
+  #   # A 2.5 minute run is unheard of, and a 2.5 hour run is unheard of
+  #   rt_vals <- rt_vals/60
+  # }
+
   rt_vals
 }